Fix command names.
