+++ lib/Drush/Commands/core/InitCommands.php
@@ -3,7 +3,7 @@ namespace Drush\Commands\core;
 
 use Drush\Log\LogLevel;
 
-class InitCommandFile extends \Robo\Tasks
+class InitCommands extends \Robo\Tasks
 {
   /**
    * Initialize local Drush configuration
