+++ tests/resources/modules/d8/woot/src/Commands/WootCommands.php
@@ -1,5 +1,5 @@
 <?php
-namespace Drupal\woot\Command;
+namespace Drupal\woot\Commands;
 
 use Consolidation\OutputFormatters\StructuredData\RowsOfFields;
